# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: RALD ELIMU — African Education Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students in a school
      parameters:
        - $ref: "#/components/parameters/schoolId"
        - in: query
          name: classId
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/StudentStatus"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: enrollStudent
      tags: [students]
      summary: Enroll a new student
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollStudentBody"
      responses:
        "201":
          description: Enrolled student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student profile
      parameters:
        - $ref: "#/components/parameters/studentId"
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student profile
      parameters:
        - $ref: "#/components/parameters/studentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{studentId}/transcript:
    get:
      operationId: getStudentTranscript
      tags: [students]
      summary: Get student academic transcript
      parameters:
        - $ref: "#/components/parameters/studentId"
      responses:
        "200":
          description: Student transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentTranscript"

  # ── Staff ──────────────────────────────────────────────────────────────────
  /schools/{schoolId}/staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members in a school
      parameters:
        - $ref: "#/components/parameters/schoolId"
        - in: query
          name: role
          schema:
            $ref: "#/components/schemas/StaffRole"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListResponse"
    post:
      operationId: addStaffMember
      tags: [staff]
      summary: Add a staff member
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStaffBody"
      responses:
        "201":
          description: Added staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{staffId}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get a staff member
      parameters:
        - $ref: "#/components/parameters/staffId"
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update staff member
      parameters:
        - $ref: "#/components/parameters/staffId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

  # ── Parents ────────────────────────────────────────────────────────────────
...
  /schools/{schoolId}/subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects in a school
      parameters:
        - $ref: "#/components/parameters/schoolId"
        - in: query
          name: academicLevel
          schema:
            $ref: "#/components/schemas/AcademicLevel"
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectListResponse"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject
      parameters:
        - $ref: "#/components/parameters/subjectId"
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - $ref: "#/components/parameters/subjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectBody"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"

  # ── Timetables ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    schoolId:
      in: path
      name: schoolId
      required: true
      schema:
        type: string
    studentId:
      in: path
      name: studentId
      required: true
      schema:
        type: string
    staffId:
      in: path
      name: staffId
      required: true
      schema:
        type: string
...
    subjectId:
      in: path
      name: subjectId
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
    StudentStatus:
      type: string
      enum: [active, graduated, transferred, suspended, withdrawn]
    StaffRole:
      type: string
      enum: [teacher, administrator, support, counselor, librarian, nurse]
    AcademicLevel:
      type: string
      enum: [kindergarten, primary, secondary, university]
    AssessmentType:
      type: string
      enum: [exam, ca, assignment, quiz, project, practical]
...
    Student:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        studentNumber:
          type: string
        schoolId:
          type: string
        classId:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          nullable: true
        enrollmentDate:
          type: string
          format: date
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        status:
          $ref: "#/components/schemas/StudentStatus"
        profileImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, studentNumber, schoolId, enrollmentDate, academicLevel, status, createdAt, updatedAt]
    StudentListResponse:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [students, total, page, limit]
    EnrollStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        enrollmentDate:
          type: string
          format: date
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        classId:
          type: string
        profileImageUrl:
          type: string
      required: [firstName, lastName, academicLevel]
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        classId:
          type: string
        status:
          $ref: "#/components/schemas/StudentStatus"
        profileImageUrl:
          type: string
    TranscriptEntry:
      type: object
      properties:
        subjectName:
          type: string
        subjectCode:
          type: string
        assessmentType:
          $ref: "#/components/schemas/AssessmentType"
        score:
          type: number
        maxScore:
          type: number
        grade:
          type: string
          nullable: true
        academicYear:
          type: string
        academicTerm:
          type: string
          nullable: true
        assessmentDate:
          type: string
          format: date
          nullable: true
      required: [subjectName, subjectCode, assessmentType, score, maxScore, academicYear]
    StudentTranscript:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptEntry"
        gpa:
          type: number
          nullable: true
        totalCredits:
          type: integer
          nullable: true
      required: [student, entries]

    # ── Staff ───────────────────────────────────────────────────────────────
    StaffMember:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        staffNumber:
          type: string
        schoolId:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        subjectIds:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        joinedDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, on_leave]
        profileImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, staffNumber, schoolId, role, subjectIds, certifications, joinedDate, status, createdAt, updatedAt]
    StaffListResponse:
      type: object
      properties:
        staff:
          type: array
          items:
            $ref: "#/components/schemas/StaffMember"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [staff, total, page, limit]
    AddStaffBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        subjectIds:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        joinedDate:
          type: string
          format: date
        profileImageUrl:
          type: string
      required: [firstName, lastName, role]
    UpdateStaffBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        subjectIds:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, on_leave]
        profileImageUrl:
          type: string

    # ── Parent ──────────────────────────────────────────────────────────────
...
    Subject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        schoolId:
          type: string
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        description:
          type: string
          nullable: true
        creditUnits:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, code, schoolId, academicLevel, createdAt, updatedAt]
    SubjectListResponse:
      type: object
      properties:
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        total:
          type: integer
      required: [subjects, total]
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        description:
          type: string
        creditUnits:
          type: integer
      required: [name, code, academicLevel]
    UpdateSubjectBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        creditUnits:
          type: integer

    # ── Timetable ────────────────────────────────────────────────────────────
...